# bench/vulnerable-api/openapi/vAPI.yaml
openapi: 3.0.1
info:
  title: Vulnerable API
  description: Vulnerable API
  version: '0.1'
paths:
  /:
    get:
      description: Return API status
      x-openapi-router-controller: vAPI
      operationId: get_root
      responses:
        '200':
          description: API status results
          content:
            'application/json':
              schema:
                type: object
              examples: {}
  /tokens:
    post:
      description: Authenticate
      x-openapi-router-controller: vAPI
      operationId: get_token
      requestBody:
        content:
          'application/json':
            schema:
              type: object
              properties:
                password:
                  type: string
                username:
                  type: string
            examples:
              '0':
                value: |-
                  {"username": "user1",
                   "password":"pass1"}
      responses:
        '200':
          description: Authentication response
          content:
            'application/json':
              schema:
                type: string
              examples: {}
        '400':
          description: Invalid input
          content:
            'application/problem+json':
              schema:
                type: string
        '401':
          description: Unauthorized
          content:
            'application/json':
              schema:
                type: string
              examples:
                '0':
                  value: |-
                    {
                      "error": {
                        "message": "username  not found"
                      }
                    }
  /user/{user}:
    get:
      x-openapi-router-controller: vAPI
      operationId: get_user
      parameters:
        - in: path
          name: user
          required: true
          schema: 
            type: string
          description: User account
        - name: X-Auth-Token
          in: header
          schema:
            type: string
          example: THISISATOKEN
      responses:
        '200':
          description: Auto generated using Swagger Inspector
          content:
            'application/json':
              schema:
                type: object
                properties:
                  response:
                    type: object
                    properties:
                      error:
                        type: object
                        properties:
                          message:
                            type: string
              examples:
                '0':
                  value: >-
                    {"response": {"error": {"message": "token id THISISATOKEN
                    not found"}}}
        '400':
          description: Invalid input
          content:
            'application/problem+json':
              schema:
                type: string
        '401':
          description: Unauthenticated
          content:
            'application/json':
              schema:
                type: string
        '403':
          description: Unauthorized
          content:
            'application/json':
              schema:
                type: string
        '404':
          description: Not found
          content:
            'application/json':
              schema:
                type: string
  /user:
    post:
      description: Auto generated using Swagger Inspector
      x-openapi-router-controller: vAPI
      operationId: create_user
      parameters:
        - name: X-Auth-Token
          in: header
          schema:
            type: string
          example: THISISATOKEN
      requestBody:
        content:
          'application/json':
            schema:
              type: object
              properties:
                password:
                  type: string
                username:
                  type: string
            examples:
              '0':
                value: "{\"username\": \"USERNAME\",\"password\": \"PASSWORD\"}"
      responses:
        '200':
          description: Auto generated using Swagger Inspector
          content:
            'application/json':
              schema:
                type: object
                properties:
                  response:
                    type: object
                    properties:
                      error:
                        type: object
                        properties:
                          message:
                            type: string
              examples:
                '0':
                  value: >-
                    {"response": {"error": {"message": "must provide valid admin token"}}}
        '400':
          description: Invalid input
          content:
            'application/json':
              schema:
                type: string
        '401':
          description: Unauthenticated
          content:
            'application/json':
              schema:
                type: string
        '403':
          description: Unauthorized
          content:
            'application/json':
              schema:
                type: string
  /uptime:
    get:
      x-openapi-router-controller: vAPI
      operationId: display_uptime
      description: Auto generated using Swagger Inspector
      responses:
        '200':
          description: Auto generated using Swagger Inspector
          content:
            'application/json':
              schema:
                type: string
              examples: {}
  /uptime/{flag}:
    get:
      x-openapi-router-controller: vAPI
      operationId: display_uptime_flag
      parameters:
        - in: path
          name: flag
          required: true
          schema: 
            type: string
      description: Flag for uptime
      responses:
        '200':
          description: Auto generated using Swagger Inspector
          content:
            'application/json':
              schema:
                type: string
              examples: {}
        '404':
          description: Not found
          content:
            'application/json':
              schema:
                type: string
  /widget:
    post:
      description: Make widget reservation
      x-openapi-router-controller: vAPI
      operationId: create_widget_reservation
      parameters:
        - name: X-Auth-Token
          in: header
          schema:
            type: string
          example: THISISATOKEN
      requestBody:
        content:
          'application/json':
            schema:
              type: object
              properties:
                name:
                  type: string
            examples:
              '0':
                value: |-
                  {"name": "widgetname"}
      responses:
        '200':
          description: Successful widget reservation
          content:
            'application/json':
              schema:
                type: string
              examples: {}
        '400':
          description: Invalid input
          content:
            'application/problem+json':
              schema:
                type: string
        '401':
          description: Unauthenticated
          content:
            'application/json':
              schema:
                type: string
        '403':
          description: Unauthorized
          content:
            'application/json':
              schema:
                type: string
  /eval:
    get:
      description: Evaluate a string
      x-openapi-router-controller: vAPI
      operationId: evaluate_str
      parameters:
        - in: header
          name: X-Auth-Token
          schema:
            type: string
          example: THISISATOKEN
        - in: query
          name: s
          schema:
            type: string
      responses:
        '200':
          description: Successful command
          content:
            'application/json':
              schema:
                type: string
  /search:
    post:
      description: Search for a user
      x-openapi-router-controller: vAPI
      operationId: search
      requestBody:
        content:
          'application/xml':
            schema:
              type: object
              properties:
                user:
                  type: string
            examples:
              '0':
                value: |-
                  <?xml version="1.0" encoding="UTF-8" ?>
                  <root>
                    <user>username</user>
                  </root>
      responses:
        '200':
          description: Search response
          content:
            'application/json':
              schema:
                type: string